% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Plot a country's population in a time series}
\usage{
CountryPopulation(name)
}
\arguments{
\item{name}{Desired country's name}
}
\value{
A plot of the country's population over time
}
\description{
This function takes a country name as an input and graphs its
population from 1950-2020. Countries not within the dataset return
an error.
}
\examples{
CountryPopulation("Mexico")
}
