% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World_Cup}
\format{
A data frame with 22 observations with 4 columns.
\describe{
     \item{WorldCup}{Identification of world cup with year and host}
     \item{Matches}{Number of matches played}
     \item{Totalattendance}{Total people attended}
     \item{Averageattendance}{Average per match attendance}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
This dataset contains the number of matches, attendance, and average attendance for
every Fifa World Cup. The data was acquired from Wikipedia
}
\keyword{datasets}
